+++ src/client/networking/mod.rs
@@ -1,5 +1,4 @@
 pub mod systems;
-pub mod messages;
 
 use crate::prelude::*;
