+++ src/client/remote_player/events.rs
@@ -1,11 +1,5 @@
 use crate::prelude::*;
 
-#[derive(Serialize, Deserialize)]
-pub struct PlayerState {
-    pub position: Vec3,
-    pub rotation: Quat,
-}
-
 #[derive(Event)]
 pub struct RemotePlayerSpawnedEvent {
     pub client_id: ClientId,
@@ -18,5 +12,5 @@ pub struct RemotePlayerDespawnedEvent {
 
 #[derive(Event)]
 pub struct RemotePlayerSyncEvent {
-    pub players: HashMap<ClientId, PlayerState>,
+    pub players: HashMap<ClientId, lib::PlayerState>,
 }
